Add a bunch of extra .pc files for additional functionality
This is done in a bit of a brute-force method at the moment.  I have .pc.in
templates for all possible files.  Then I check the underlying cairo-XXX.pc
files and if those exist, I generate the corresponding cairomm-XXX.pc file.
It's not the prettiest solution, but it seems to work.
So now, if you want to just use base cairo functionality, check for cairomm-1.0,
but if you want to use other functionality that requires that you link your
application against other libraries (e.g. freetype, xlib, etc), you should check
for cairomm-ft-1.0, cairomm-xlib-1.0, etc.