Update and Overhaul the Visual Studio 2008 Projects
Move all the projects to MSVC_Net2008/, and add property sheets for the
projects so that commonly-used items can be consolidated and referred from
the projects, which will help simplify future maintenance.  Also remove
items from the projects that aren't really needed, as they add quite a bit
of clutter.  Add the newly-added source files to the cairomm project as
well, as the cairomm project has bit rotted a little.

https://bugs.freedesktop.org/show_bug.cgi?id=84030