cairomm/exception.h: Allow Build on Visual Studio 2013
Visual Studio 2013 has one catch for its C++-11 support: It does
not support noexcept, but has its own _NOEXCEPT for the same purpose.
Add a define for it, while defining the _ALLOW_KEYWORD_MACROS as newer
Visual Studio Versions do not allow one to define known keywords, even if
it is not supported, by default.